Switch from using `admin-ajax.php` to calling the current page with additional query args.
`admin-ajax.php` uses a different context from normal front page calls. If we want widgets to be rendered as if they were on the original
page, it's much easier to send the AJAX call call the current URL with our query string appended and then interrupt the rendering at
`template_redirect`. Otherwise we would have to save and restore the main query / globals which is painful in WordPress.